+++ lib/byebug/processors/command_processor.rb
@@ -5,7 +5,7 @@ module Byebug
   # Processes commands in regular mode
   #
   class CommandProcessor < Processor
-    attr_reader :display
+    attr_reader :display, :state
 
     def initialize(interface = LocalInterface.new)
       super(interface)
@@ -117,35 +117,29 @@ module Byebug
     # @return List of commands acceptable to run bound to the current state
     #
     def always_run(context, file, line, run_level)
-      cmds = Command.commands
-
-      state = Byebug::RegularState.new(cmds,
-                                       context,
-                                       @display,
-                                       file,
-                                       @interface,
-                                       line)
+      @state = Byebug::RegularState.new(Command.commands,
+                                        context,
+                                        @display,
+                                        file,
+                                        @interface,
+                                        line)
 
       # Change default when in irb or code included in command line
       Setting[:autolist] = false if ['(irb)', '-e'].include?(file)
 
       # Bind commands to the current state.
-      commands = cmds.map do |cmd_class|
-        cmd = cmd_class.new(state)
-        cmd.execute if cmd.class.always_run >= run_level
-        cmd
+      Command.commands.each do |cmd|
+        cmd.new(@state).execute if cmd.always_run >= run_level
       end
-
-      [state, commands]
     end
 
     #
     # Handle byebug commands.
     #
     def process_commands(context, file, line)
-      state, commands = preloop(context, file, line)
+      preloop(context, file, line)
 
-      repl(state, commands, context)
+      repl(context)
 
       postloop
     end
@@ -153,8 +147,8 @@ module Byebug
     #
     # Main byebug's REPL
     #
-    def repl(state, commands, context)
-      until state.proceed?
+    def repl(context)
+      until @state.proceed?
         cmd = @interface.read_command(prompt(context))
         return unless cmd
 
@@ -162,27 +156,30 @@ module Byebug
 
         cmd.empty? ? cmd = @last_cmd : @last_cmd = cmd
 
-        one_cmd(commands, context, cmd)
+        one_cmd(context, cmd)
       end
     end
 
     #
     # Autoevals a single command
     #
-    def one_unknown_cmd(commands, input)
+    def one_unknown_cmd(input)
       unless Setting[:autoeval]
         return errmsg("Unknown command: \"#{input}\". Try \"help\"")
       end
 
-      commands.find { |c| c.is_a?(EvalCommand) }.execute
+      eval_cmd = EvalCommand.new(@state)
+      eval_cmd.match(input)
+      eval_cmd.execute
     end
 
     #
     # Executes a single byebug command
     #
-    def one_cmd(commands, context, input)
-      cmd = commands.find { |c| c.match(input) }
-      return one_unknown_cmd(commands, input) unless cmd
+    def one_cmd(context, input)
+      cmd = match_cmd(input)
+
+      return one_unknown_cmd(input) unless cmd
 
       if context.dead? && !cmd.class.allow_in_post_mortem
         return errmsg('Command unavailable in post mortem mode.')
@@ -192,21 +189,28 @@ module Byebug
     end
 
     #
+    # Finds a matches the command matching the input
+    #
+    def match_cmd(input)
+      Command.commands.each do |c|
+        cmd = c.new(@state)
+        return cmd if cmd.match(input)
+      end
+
+      nil
+    end
+
+    #
     # Tasks to do before processor loop.
     #
     def preloop(context, file, line)
-      state, commands = always_run(context, file, line, 1)
-
-      thread_state = Setting[:testing] ? state : nil
-      Thread.current.thread_variable_set('state', thread_state)
+      always_run(context, file, line, 1)
 
       puts 'The program finished.' if program_just_finished?(context)
 
-      puts(state.location) if Setting[:autolist] == 0
+      puts(@state.location) if Setting[:autolist] == 0
 
       @interface.history.restore if Setting[:autosave]
-
-      [state, commands]
     end
 
     #