+++ test/commands/show_test.rb
@@ -14,7 +14,7 @@ module Byebug
     end
 
     settings = %i(autoeval autolist autosave basename fullpath post_mortem
-                  stack_on_error testing)
+                  stack_on_error)
 
     settings.each do |set|
       define_method(:"test_show_#{set}_shows_disabled_bool_setting_#{set}") do